+++ src/core/object/gameobject_manager.hh
@@ -17,7 +17,7 @@ class GameObjectManager final
     static eastl::fixed_vector<GameObject *, MAX_GAME_OBJECTS> m_activeGameObjects;
     static eastl::fixed_vector<GameObject *, MAX_GAME_OBJECTS> m_renderableGameObjects;
 
-    static int8_t GetFreeIndex(void);
+    static int16_t GetFreeIndex(void);
 
 public:
     static GameObject *CreateGameObject(const char *name, psyqo::Vec3 pos, GameObjectRotation rotation, GameObjectTag tag = GameObjectTag::NONE);
